types: change len to intmax_t
Using a signed type for the length isn't great, but it is what emacs
uses to represent integers.  Also, a value of less-than-0 will error out
`extract_string()`.  Besides, the unsigned `size` member is what's
ultimately used for memory allocations.

+++ src/types.h
@@ -11,7 +11,7 @@
 
 struct string {
     char *str;
-    size_t len;
+    intmax_t len;
     size_t size;
 };
